+++ polardbx-20200202/core/Models/ModifyDBInstanceClassResponseBody.cs
@@ -29,6 +29,10 @@ namespace AlibabaCloud.SDK.Polardbx20200202.Models
         [Validation(Required=false)]
         public string RequestId { get; set; }
 
+        [NameInMap("TaskId")]
+        [Validation(Required=false)]
+        public int? TaskId { get; set; }
+
     }
 
 }